Updated Physics engine
Updated physics engine, now collision should work correctly. I'm still not sure if this will be the final version because it can have problem if multiple entity get caught in the player AABB. In this case the entity with the lower index in the entity vector will be calculated first, but it may not be the first entity colliding with the player. This could create some unexpected situation but for now it's good as it is.