Joystick support
This commit added a basic gamepad support. Default keys (in inc/gameEngine.hpp) are trash because I currently can't test this code with a real gamepad, I will update them as soon as I can test with a joystick.
In the future I will create a dedicated menu to change the key settings ingame.