Mouse interagible MainMenu
Added mouse intragibility to the MainMenu. Now I have to do it for the SettingMenu but it will be a lot of work because there are a lot of texts.
I also added an aspect ratio check for the level editor, this way I can tell to the user that it's designed for 16:9 only. In the future I should probably just fix the viewport to make it scale with different resolution but for now it's good this way.